Fix floating-point math to actually use floating point numbers

+++ melee.cpp
@@ -592,7 +592,7 @@ int player::roll_cut_damage(monster *z, bool crit)
  if (z_armor_cut < 0)
   z_armor_cut = 0;
 
- int ret = weapon.damage_cut() - z_armor_cut;
+ double ret = weapon.damage_cut() - z_armor_cut;
 
  if (unarmed_attack() && !wearing_something_on(bp_hands)) {
   if (has_trait(PF_CLAWS))
@@ -613,7 +613,7 @@ int player::roll_cut_damage(monster *z, bool crit)
   ret *= double( 0.92 + 0.04 * skillLevel("cutting") );
 
  if (crit)
-  ret *= double( 1.0 + double(skillLevel("cutting") / 12) );
+  ret *= 1.0 + (skillLevel("cutting") / 12.0);
 
  return ret;
 }
@@ -654,7 +654,7 @@ int player::roll_stab_damage(monster *z, bool crit)
   return 0; // No negative stabbing!
 
  if (crit) {
-  int multiplier = double( 1.0 + double(skillLevel("stabbing") / 5) );
+  double multiplier = 1.0 + (skillLevel("stabbing") / 5.0);
   if (multiplier > 2.5)
    multiplier = 2.5;
   ret *= multiplier;
